Create methods for Surface moved
