+++ Physical_Engine/Physical_Engine.csproj
@@ -63,7 +63,6 @@
     <Compile Include="Create\Wall.cs" />
     <Compile Include="Create\Layer.cs" />
     <Compile Include="Create\Material.cs" />
-    <Compile Include="Create\Surface.cs" />
     <Compile Include="Modify\AddMaterialProperties.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Query\Construction.cs" />
